+++ packages/core/icon-file-type/src/components/__tests__/unit/server-side.js
@@ -6,8 +6,8 @@ import React from 'react';
 import { getExamplesFor } from '@atlaskit/build-utils/getExamples';
 import ReactDOMServer from 'react-dom/server';
 
-test('Icon server side rendering', async () => {
-  (await getExamplesFor('icon')).forEach(examples => {
+test('Icon file type server side rendering', async () => {
+  (await getExamplesFor('icon-file-type')).forEach(examples => {
     // $StringLitteral
     const Example = require(examples.filePath).default; // eslint-disable-line import/no-dynamic-require
     expect(() => ReactDOMServer.renderToString(<Example />)).not.toThrowError();