+++ packages/editor/editor-core/src/create-editor/create-plugins-list.ts
@@ -49,6 +49,7 @@ import {
   editorDisabledPlugin,
   indentationPlugin,
   annotationPlugin,
+  compositionPlugin,
 } from '../plugins';
 
 /**
@@ -231,6 +232,10 @@ export default function createPluginsList(
 
   if (props.appearance !== 'mobile') {
     plugins.push(quickInsertPlugin);
+  }
+
+  if (props.appearance === 'mobile') {
+    plugins.push(compositionPlugin);
   }
 
   if (props.appearance === 'message') {